Fixed an issue where sometimes the shadows wouldn't be computed correctly from inside a polygon.
Sometimes, when a token was placed inside a a structure surrounded by
vision obstructions, the shadows wouldn't be computed correctly and
everything would be black. To fix this, we now use only one object that
cuts the base fog layer (fill fog). The extension now also automatically
enables fill fog upon drawing if it is not enabled already.